Split Image into SimpleImage
